Move service cmds to a separate file
